[xla] Saves original value information in fusion instructions
Currently TPU logging only prints out the information of a fusion instruction, but not the intermediate values in fused comptation. compare_hlo_output uses origin attribute as a key for comparing instructions, and as a result it cannot compare original values of fusion instructions properly, since original values are saved in the fused root, but not the instruction itself. This is a workaround.

PiperOrigin-RevId: 677923103